rentify/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
click==8.1.7
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
django-tastypie==0.14.6
django-tastypie-swagger==0.1.4
djangorestframework==3.14.0
exceptiongroup==1.2.0
fastapi==0.108.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
idna==3.6
Markdown==3.5.1
Pillow==10.1.0
promise==2.3
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.32.0.post1
text-unidecode==1.3
typing_extensions==4.9.0
uvicorn==0.25.0
